Add files via upload
In this project, as a sales data analyst, I was tasked with analyzing sales data to provide insights on the performance of a company's products and customers. 

Through my analysis, I was able to determine:

the best month for sales in a specific year and the product that was most successful during that period. Using RFM analysis, 
I identified the best customer for the company, which allowed them to focus on retaining high-value customers and improving their overall customer experience. 
Additionally, by identifying which products were most often sold together, I helped the company create targeted product bundles and promotions that boosted sales.
Finally, by pinpointing the city with the highest number of sales in a specific country, I provided the company with valuable geographic insights that could inform their expansion plans and marketing tactics. 

Overall, my work as a sales data analyst added immense value to the client by providing crucial insights that informed their business decisions and helped them achieve their objectives.